tmp mp4 file removed while deleting original mp4

+++ cron/delete-existing-mp4-videos.py
@@ -22,7 +22,11 @@ rows = cur.fetchall()
 for row in rows:
     ogv_video_path = MEDIA_ROOT + 'videos/' + str(row[8]) + '/' + str(row[1]) + '/' + row[5]
     mp4_video_path, file_extension = os.path.splitext(ogv_video_path)
+    mp4_tmp_video_path = mp4_video_path + '-tmp.mp4'
     mp4_video_path = mp4_video_path + '.mp4'
-    if os.path.isfile(mp4_video_path):
+    if os.path.isfile(mp4_video_path) or os.path.isfile(mp4_tmp_video_path):
         print row[5]
-        os.remove(mp4_video_path)
+        if os.path.isfile(mp4_video_path):
+            os.remove(mp4_video_path)
+        if os.path.isfile(mp4_tmp_video_path):
+            os.remove(mp4_tmp_video_path)
